Working towards replacing old context code with new (2).
